update the fused.h
