+++ lib/runtime/src/ops/fused.h
@@ -29,16 +29,8 @@
 #include "kernel/transpose_kernels.h"
 #include "config.h"
 
+
 namespace FlexFlow {
-
-enum SourceType { 
-    SOURCE_NONE,
-    SOURCE_INPUT, 
-    SOURCE_WEIGHT,
-     SOURCE_OUTPUT, 
-  }; 
-
-class FusedOP;
 
 using AllDevice = variant<BatchMatmulPerDeviceState, BatchNormPerDeviceState, CastPerDeviceState, CombinePerDeviceState, ConcatPerDeviceState,  Conv2DPerDeviceState, DropoutPerDeviceState, ElementBinaryPerDeviceState, ElementUnaryPerDeviceState, EmbeddingPerDeviceState, FlatPerDeviceState, GatherPerDeviceState, GroupByPerDeviceState, LayerNormPerDeviceState, LinearPerDeviceState, RepartitionPerDeviceState, Pool2DPerDeviceState, ReducePerDeviceState, ReshapePerDeviceState, SoftmaxPerDeviceState, TopKPerDeviceState, TransposePerDeviceState>;//this can update to hold all device state
 
@@ -46,34 +38,10 @@ struct FusedPerDeviceOpState{
     FusedOp fused_op;
     int numOperators;
     //maybe we can all xxxPerdeviceOpState here，use Variant to store all device 
-
+    AllDevice all_device;
 };
 
-struct FusedOP : public OP {
-    int numOperators; 
-    int numInputs;
-    int numWeights;
-    int numOutputs;
-    
-    FFIterationConfig iter_config; 
-    int op_num_inputs[MAX_NUM_FUSED_OPERATORS]; 
-    int op_num_weights[MAX_NUM_FUSED_OPERATORS]; 
-    int op_num_outputs[MAX_NUM_FUSED_OPERATORS]; 
-    OperatorType op_op_type[MAX_NUM_FUSED_OPERATORS]; 
-    SourceType op_input_source[MAX_NUM_FUSED_TENSORS]; 
-    SourceType op_weight_source[MAX_NUM_FUSED_TENSORS]; 
-    SourceType op_output_source[MAX_NUM_FUSED_TENSORS];
-    DataType input_data_types[MAX_NUM_INPUTS]; 
-    DataType weight_data_types[MAX_NUM_WEIGHTS]; 
-    DataType output_data_types[MAX_NUM_OUTPUTS];
-    int op_input_idx[MAX_NUM_FUSED_TENSORS]; */
-    int op_weight_idx[MAX_NUM_FUSED_TENSORS]; */
-    int op_output_idx[MAX_NUM_FUSED_TENSORS]; */
-    Op *operators[MAX_NUM_FUSED_OPERATORS]; */
-    FusedPerDeviceOpState fused_meta[MAX_NUM_WORKERS]; 
-    int numOperators; 
 
-};
 
 template <>
 void register_task<FUSEDOP_INIT_TASK_ID>();
